requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
aiomysql==0.2.0
pydantic==2.8.2
python-dotenv==1.0.1
greenlet==3.0.3
